+++ packages/preview-middleware/src/base/flp.ts
@@ -56,6 +56,13 @@ type OnChangeRequestHandler = (
     logger: Logger
 ) => Promise<void>;
 
+type Ui5Version = {
+    major: number;
+    minor: number;
+    patch: number;
+    label?: string;
+};
+
 /**
  * Class handling preview of a sandbox FLP.
  */
@@ -222,7 +229,7 @@ export class FlpSandbox {
         if (ui5Version.major === 1 && ui5Version.minor <= 71) {
             this.removeAsyncHintsRequests();
         }
-        return render(this.getSandboxTemplate(ui5Version.major), config);
+        return render(this.getSandboxTemplate(ui5Version), config);
     }
 
     /**
@@ -376,7 +383,7 @@ export class FlpSandbox {
                 req.headers.host,
                 'ui5-patched-router' in req ? req['ui5-patched-router']?.baseUrl : undefined
             );
-            const html = render(this.getSandboxTemplate(ui5Version.major), this.templateConfig);
+            const html = render(this.getSandboxTemplate(ui5Version), this.templateConfig);
             this.sendResponse(res, 'text/html', 200, html);
         }
     }
@@ -415,7 +422,7 @@ export class FlpSandbox {
         protocol: Request['protocol'],
         host: Request['headers']['host'],
         baseUrl: string = ''
-    ): Promise<{ major: number; minor: number }> {
+    ): Promise<Ui5Version> {
         let version: string | undefined;
         if (!host) {
             this.logger.error('Unable to fetch UI5 version: No host found in request header.');
@@ -434,25 +441,30 @@ export class FlpSandbox {
             this.logger.error('Could not get UI5 version of application. Using 1.121.0 as fallback.');
             version = '1.121.0';
         }
-        const [major, minor] = version.split('.').map((versionPart) => parseInt(versionPart, 10));
+        const [major, minor, patch] = version.split('.').map((versionPart) => parseInt(versionPart, 10));
+        const label = version.split(/-(.*)/s)?.[1];
         return {
             major,
-            minor
+            minor,
+            patch,
+            label
         };
     }
 
     /**
      * Read the sandbox template file based on the given UI5 version.
      *
-     * @param ui5MajorVersion - the major version of UI5
+     * @param ui5Version - the UI5 version
      * @returns the template for the sandbox HTML file
      */
-    private getSandboxTemplate(ui5MajorVersion: number): string {
-        this.logger.info(`Using sandbox template for UI5 major version ${ui5MajorVersion}.`);
-        return readFileSync(
-            join(__dirname, `../../templates/flp/sandbox${ui5MajorVersion === 1 ? '' : ui5MajorVersion}.html`),
-            'utf-8'
+    private getSandboxTemplate(ui5Version: Ui5Version): string {
+        this.logger.info(
+            `Using sandbox template for UI5 version ${ui5Version.major}.${ui5Version.minor}.${ui5Version.patch}${
+                ui5Version.label ? `-${ui5Version.label}` : ''
+            }.`
         );
+        const filePrefix = ui5Version.major > 1 || ui5Version.label?.includes('legacy-free') ? '2' : '';
+        return readFileSync(join(__dirname, `../../templates/flp/sandbox${filePrefix}.html`), 'utf-8');
     }
 
     /**